[WIP] break out a few contexts, start refactoring fast index modules
